refactor(extension): improve package
