+++ example/compose/src/main/java/com/neoutils/highlight/example/compose/example/CaptureGroupsExample.kt
@@ -1,15 +1,12 @@
-package com.neoutils.highlight.example.compose.exemple
+package com.neoutils.highlight.example.compose.example
 
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.remember
 import androidx.compose.ui.tooling.preview.Preview
 import com.neoutils.highlight.compose.extension.toAnnotatedString
-import com.neoutils.highlight.core.Highlight
-import com.neoutils.highlight.core.scheme.TextColorScheme
-import com.neoutils.highlight.core.utils.Match
+import com.neoutils.highlight.core.ktx.highlight
 import com.neoutils.highlight.core.utils.UiColor
-import com.neoutils.highlight.extension.highlight
 
 @Composable
 fun CaptureGroupsExample() {
