+++ example/view-based/src/main/kotlin/com/neoutils/hightlight/example/view/example/BackgroundSchemeExample.kt
@@ -3,12 +3,8 @@ package com.neoutils.hightlight.example.view.example
 import android.content.Context
 import android.util.AttributeSet
 import androidx.appcompat.widget.AppCompatTextView
-import com.neoutils.highlight.core.Highlight
-import com.neoutils.highlight.core.scheme.BackgroundColorScheme
-import com.neoutils.highlight.core.scheme.TextColorScheme
-import com.neoutils.highlight.core.utils.Match
+import com.neoutils.highlight.core.ktx.highlight
 import com.neoutils.highlight.core.utils.UiColor
-import com.neoutils.highlight.extension.highlight
 import com.neoutils.highlight.view.extension.toSpannedString
 
 class BackgroundSchemeExample(
@@ -24,13 +20,13 @@ class BackgroundSchemeExample(
         text = highlight {
             backgroundColor {
                 fully(
-                    regex = "background color",
+                    regex = "color",
                     UiColor.Blue
                 )
             }
             textColor {
                 fully(
-                    regex = "background color",
+                    regex = "color",
                     UiColor.White
                 )
             }
