+++ example/view-based/src/main/kotlin/com/neoutils/hightlight/example/view/example/TextColorSchemeExample.kt
@@ -3,11 +3,11 @@ package com.neoutils.hightlight.example.view.example
 import android.content.Context
 import android.util.AttributeSet
 import androidx.appcompat.widget.AppCompatTextView
+import com.neoutils.highlight.core.ktx.highlight
 import com.neoutils.highlight.core.utils.UiColor
-import com.neoutils.highlight.extension.highlight
 import com.neoutils.highlight.view.extension.toSpannedString
 
-class TextColorSchemeExample(
+class TextColorSchemeExample @JvmOverloads constructor(
     context: Context,
     attr: AttributeSet? = null,
 ) : AppCompatTextView(context, attr) {
@@ -17,7 +17,7 @@ class TextColorSchemeExample(
         text = highlight {
             textColor {
                 fully(
-                    regex = "foreground color",
+                    regex = "color",
                     value = UiColor.Blue
                 )
             }
